backend/requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.2.0
freezegun==1.2.2
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
packaging==23.1
pg8000==1.30.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-env==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
